modify redirection for N|Solid version resolution
fixes https://github.com/nodesource/heroku-buildpack-nsolid/issues/4

Not sure what the exact problem was - perhaps the `jq` output got so large
that it somehow effected how the piping worked?  An easier explanation would
be a change to `bash` itself, but not sure how to verify that.

In any case, this seems to work.